Port to .NET Core
