+++ Histacom2/OS/Win95/Win95Apps/IE3Sites/Website12padams1998.resx
