+++ Histacom2/OS/Win95/Windows95.resx
