+++ Histacom2/OS/Win95/Windows95TaskbarItem.resx
