+++ Histacom2/OS/Win98/Windows98.resx
