+++ Histacom2/Properties/AssemblyInfo.cs
@@ -3,18 +3,6 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Histacom2")]
-[assembly: AssemblyDescription("Revival of the original hacking simulator!")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Michael Webb")]
-[assembly: AssemblyProduct("Histacom2")]
-[assembly: AssemblyCopyright("Copyright © Michael Webb and contributors, 2019.")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
 // Setting ComVisible to false makes the types in this assembly not visible 
 // to COM components.  If you need to access a type in this assembly from 
 // COM, set the ComVisible attribute to true on that type.
@@ -22,18 +10,3 @@ using System.Runtime.InteropServices;
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("b958aa5f-8aed-456c-9c07-9e4c6c027042")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.0.3")]
-[assembly: AssemblyFileVersion("2.0.3")]
-[assembly: NeutralResourcesLanguage("en-US")]
-
